Add brc keymap for ErgoDox.
I think this is around the time in the commit log that I had done a "git pull"
and then compiled this code.  The mtimes on my files were over two years old,
with Jan-20-2021 timestamps, and when I did `git checkout "master@{2 years ago}",
which I thought should have brought me to March-24-2021, it dropped me at
exactly Jan 20th, 2021 (I think that's because of what's in my local reflog--if
you clone the upstream repo fresh and try that command, it doesn't work because
there is no reflog history).

Anywho, YMMV if you ever use this again, future-Brett. :-)